+++ test/e2e/suite_test.go
@@ -46,7 +46,7 @@ const (
 	defaultReportDir = "/workspace/_artifacts"
 )
 
-var blobfuseDriver *blobfuse.Driver
+var blobDriver *blob.Driver
 
 type testCmd struct {
 	command  string
@@ -96,10 +96,10 @@ var _ = ginkgo.BeforeSuite(func() {
 	}
 	nodeid := os.Getenv("nodeid")
 	kubeconfig := os.Getenv(kubeconfigEnvVar)
-	blobfuseDriver = blobfuse.NewDriver(nodeid)
+	blobDriver = blob.NewDriver(nodeid)
 	go func() {
 		os.Setenv("AZURE_CREDENTIAL_FILE", credentials.TempAzureCredentialFilePath)
-		blobfuseDriver.Run(fmt.Sprintf("unix:///tmp/csi-%s.sock", uuid.NewUUID().String()), kubeconfig)
+		blobDriver.Run(fmt.Sprintf("unix:///tmp/csi-%s.sock", uuid.NewUUID().String()), kubeconfig)
 	}()
 })
