+++ test/e2e/testsuites/dynamically_provisioned_reclaim_policy_tester.go
@@ -29,7 +29,7 @@ import (
 type DynamicallyProvisionedReclaimPolicyTest struct {
 	CSIDriver              driver.DynamicPVTestDriver
 	Volumes                []VolumeDetails
-	Blobfuse               *blobfuse.Driver
+	Driver                 *blob.Driver
 	StorageClassParameters map[string]string
 }
 
@@ -44,7 +44,7 @@ func (t *DynamicallyProvisionedReclaimPolicyTest) Run(client clientset.Interface
 		if tpvc.ReclaimPolicy() == v1.PersistentVolumeReclaimRetain {
 			tpvc.WaitForPersistentVolumePhase(v1.VolumeReleased)
 			tpvc.DeleteBoundPersistentVolume()
-			tpvc.DeleteBackingVolume(t.Blobfuse)
+			tpvc.DeleteBackingVolume(t.Driver)
 		}
 	}
 }
